@@ -1,5 +1,7 @@
 const React = require('react');
 const ReactDOM = require('react-dom');
 const Board = require('./Board');
+const Deck = require('./Deck');
 
-ReactDOM.render(<Board size={12}/>, document.getElementById('memory'));
+const deck = new Deck(12);
+ReactDOM.render(<Board deck={deck}/>, document.getElementById('memory'));
